Updated run.sh to be more portable

+++ client/run.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
+SETUP_FILE=/var/lightberry.setup
+
 if [ -f "$1" ]; then
-	source $1
+	. $1
 fi
 
 if [ ! -d "$DEVICE_CLASS" ] || [ ! -f "$DEVICE_CLASS/device.py" ]; then
@@ -8,6 +10,9 @@ if [ ! -d "$DEVICE_CLASS" ] || [ ! -f "$DEVICE_CLASS/device.py" ]; then
 	exit 1
 fi
 
-pip3 install -r requirements.txt -r $DEVICE_CLASS/requirements.txt
+if [ ! -f "$SETUP_FILE" ]; then
+	pip3 install -r requirements.txt -r $DEVICE_CLASS/requirements.txt
+	touch $SETUP_FILE
+fi
 
-python3 /usr/src/app/main.py
+python3 $CLIENT_DIR/main.py
